+++ test/test_helper.exs
@@ -3,5 +3,4 @@
 |> Path.wildcard()
 |> Enum.each(&File.rm_rf/1)
 
-ExUnit.start()
-Application.ensure_all_started(:bypass)
+ExUnit.start(capture_log: true)